requirements.txt
pandas>=2.0
openpyxl>=3.1
click>=8.1
pyyaml>=6.0
python-dotenv>=1.0
pytest>=7.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0
motor>=3.3.0
together>=1.0.0
